Shuffling files.
