+++ SDK/plugin.h
@@ -6,7 +6,7 @@
 //----------------------------------------------------------
 
 #include "plugincommon.h"
-#include "amx/amx.h"
+#include "amx.h"
 
 //----------------------------------------------------------
 // EOF
